Add files via upload
insert_sentiment_analyze_with_bert.sql
This SQL script is designed to insert the results of the BERT sentiment analysis into the bert_sentiment_scores table. It ensures that the analyzed data, including the sentiment label (POSITIVE or NEGATIVE) and the confidence score (sentiment_score), is accurately and efficiently stored in the database. This script is a crucial part of the data pipeline, as it prepares the BERT results for comparison with other analyses and for subsequent visualization.